pyproject.toml
[project]
name = "gemini-demo"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizuki", email = "mbaba@kanolab.net" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.1",
    "IPython>=8.18.1",
    "pandas>=2.1.4",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/foo"]
